styled the actual page that displays the articles
since data gathering is taking a bit longer, the page that will be served when accessing articles was styled with dummy text. the current route is /changelater but will be changed to a dynamic route later on when my data collection is done. should be the last bit to complete on the frontend